changes to PR that fixes #76 - renaming ResultsUploader to ResultsWriter and moving all write methods to this class
