+++ src/main/java/com/oltpbenchmark/Results.java
@@ -19,18 +19,14 @@
 package com.oltpbenchmark;
 
 import com.oltpbenchmark.LatencyRecord.Sample;
-import com.oltpbenchmark.ThreadBench.TimeBucketIterable;
 import com.oltpbenchmark.api.TransactionType;
 import com.oltpbenchmark.util.Histogram;
-import com.oltpbenchmark.util.StringUtil;
 
-import java.io.PrintStream;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 public final class Results {
-    public static final double MILLISECONDS_FACTOR = 1e3;
 
     private final long nanoseconds;
     private final int measuredRequests;
@@ -94,121 +90,23 @@ public final class Results {
         return (double) measuredRequests / (double) nanoseconds * 1e9;
     }
 
+    public List<Sample> getLatencySamples() {
+        return latencySamples;
+    }
+
+    public long getNanoseconds() {
+        return nanoseconds;
+    }
+
+    public int getMeasuredRequests() {
+        return measuredRequests;
+    }
+
     @Override
     public String toString() {
         return "Results(nanoSeconds=" + nanoseconds + ", measuredRequests=" + measuredRequests + ") = " + requestsPerSecond() + " requests/sec";
     }
 
-    public void writeResults(int windowSizeSeconds, PrintStream out) {
-        writeResults(windowSizeSeconds, out, TransactionType.INVALID);
-    }
 
-    public void writeResults(int windowSizeSeconds, PrintStream out, TransactionType txType) {
-        String[] header = {
-                "Time (seconds)",
-                "Throughput (requests/second)",
-                "Average Latency (millisecond)",
-                "Minimum Latency (millisecond)",
-                "25th Percentile Latency (millisecond)",
-                "Median Latency (millisecond)",
-                "75th Percentile Latency (millisecond)",
-                "90th Percentile Latency (millisecond)",
-                "95th Percentile Latency (millisecond)",
-                "99th Percentile Latency (millisecond)",
-                "Maximum Latency (millisecond)",
-                "tp (req/s) scaled"
-        };
-        out.println(StringUtil.join(",", header));
-        int i = 0;
-        for (DistributionStatistics s : new TimeBucketIterable(latencySamples, windowSizeSeconds, txType)) {
-            out.printf("%d,%.3f,%.3f,%.3f,%.3f,%.3f,%.3f,%.3f,%.3f,%.3f,%.3f,%.3f\n",
-                    i * windowSizeSeconds,
-                    (double) s.getCount() / windowSizeSeconds,
-                    s.getAverage() / MILLISECONDS_FACTOR,
-                    s.getMinimum() / MILLISECONDS_FACTOR,
-                    s.get25thPercentile() / MILLISECONDS_FACTOR,
-                    s.getMedian() / MILLISECONDS_FACTOR,
-                    s.get75thPercentile() / MILLISECONDS_FACTOR,
-                    s.get90thPercentile() / MILLISECONDS_FACTOR,
-                    s.get95thPercentile() / MILLISECONDS_FACTOR,
-                    s.get99thPercentile() / MILLISECONDS_FACTOR,
-                    s.getMaximum() / MILLISECONDS_FACTOR,
-                    MILLISECONDS_FACTOR / s.getAverage());
-            i += 1;
-        }
-    }
-
-    public void writeSamples(PrintStream out) {
-        writeSamples(1, out, TransactionType.INVALID);
-    }
-
-    public void writeSamples(int windowSizeSeconds, PrintStream out, TransactionType txType) {
-        String[] header = {
-                "Time (seconds)",
-                "Requests",
-                "Throughput (requests/second)",
-                "Minimum Latency (microseconds)",
-                "25th Percentile Latency (microseconds)",
-                "Median Latency (microseconds)",
-                "Average Latency (microseconds)",
-                "75th Percentile Latency (microseconds)",
-                "90th Percentile Latency (microseconds)",
-                "95th Percentile Latency (microseconds)",
-                "99th Percentile Latency (microseconds)",
-                "Maximum Latency (microseconds)"
-        };
-        out.println(StringUtil.join(",", header));
-        int i = 0;
-        for (DistributionStatistics s : new TimeBucketIterable(latencySamples, windowSizeSeconds, txType)) {
-            out.printf("%d,%d,%.3f,%d,%d,%d,%d,%d,%d,%d,%d,%d\n",
-                    i * windowSizeSeconds,
-                    s.getCount(),
-                    (double) s.getCount() / windowSizeSeconds,
-                    (int) s.getMinimum(),
-                    (int) s.get25thPercentile(),
-                    (int) s.getMedian(),
-                    (int) s.getAverage(),
-                    (int) s.get75thPercentile(),
-                    (int) s.get90thPercentile(),
-                    (int) s.get95thPercentile(),
-                    (int) s.get99thPercentile(),
-                    (int) s.getMaximum());
-            i += 1;
-        }
-    }
-
-    public void writeRaw(List<TransactionType> activeTXTypes, PrintStream out) {
-
-        // This is needed because nanTime does not guarantee offset... we
-        // ground it (and round it) to ms from 1970-01-01 like currentTime
-        double x = ((double) System.nanoTime() / (double) 1000000000);
-        double y = ((double) System.currentTimeMillis() / (double) 1000);
-        double offset = x - y;
-
-        // long startNs = latencySamples.get(0).startNs;
-        String[] header = {
-                "Transaction Type Index",
-                "Transaction Name",
-                "Start Time (microseconds)",
-                "Latency (microseconds)",
-                "Worker Id (start number)",
-                "Phase Id (index in config file)"
-        };
-        out.println(StringUtil.join(",", header));
-        for (Sample s : latencySamples) {
-            double startUs = ((double) s.getStartNanosecond() / (double) 1000000000);
-            String[] row = {
-                    Integer.toString(s.getTransactionType()),
-                    // Important!
-                    // The TxnType offsets start at 1!
-                    activeTXTypes.get(s.getTransactionType() - 1).getName(),
-                    String.format("%10.6f", startUs - offset),
-                    Integer.toString(s.getLatencyMicrosecond()),
-                    Integer.toString(s.getWorkerId()),
-                    Integer.toString(s.getPhaseId()),
-            };
-            out.println(StringUtil.join(",", row));
-        }
-    }
 
 }
